Corrected goofy bug related to symlinks & directories.

+++ autoinstall.sh
@@ -3,7 +3,7 @@
 ln -s ~/dotfiles/gitignore ~/.gitignore
 ln -s ~/dotfiles/ackrc ~/.ackrc
 ln -s ~/dotfiles/gitconfig ~/.gitconfig
-ln -s ~/dotfiles/oh-my-zsh ~/.oh-my-zsh
+[ ! -e ~/.oh-my-zsh ] && ln -s ~/dotfiles/oh-my-zsh ~/.oh-my-zsh
 ln -s ~/dotfiles/zshrc ~/.zshrc
 ln -s ~/dotfiles/vimrc ~/.vimrc
 ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
